+++ tests/unit/models/test_user.py
@@ -12,7 +12,6 @@ class TestUser(UnitTest):
         reddit = Reddit(
             client_id="dummy",
             client_secret="dummy",
-            praw8_raise_exception_on_me=True,
             user_agent="dummy",
         )
         reddit._core._requestor._http = None
@@ -21,7 +20,3 @@ class TestUser(UnitTest):
         user = User(reddit)
         with pytest.raises(ReadOnlyException):
             user.me()
-
-    def test_me__in_read_only_mode__deprecated(self, reddit):
-        assert reddit.read_only
-        assert User(reddit).me() is None