+++ CHANGES.rst
@@ -6,6 +6,11 @@ PRAW follows `semantic versioning <https://semver.org/>`_.
 Unreleased
 ----------
 
+**Changed**
+
+- Change ``Reddit.user.me`` to raise :class:`.ReadOnlyException` when called in
+  :attr:`.read_only` mode.
+
 **Removed**
 
 - Remove ``APIException`` class.
@@ -250,9 +255,9 @@ Unreleased
   an instance of :class:`.Reddit`, and via the ``PRAW_REFRESH_TOKEN`` environment
   variable. To be prepared for PRAW 8, use the new :class:`.Reddit` keyword argument
   ``token_manager``. See :ref:`refresh_token` in PRAW's documentation for an example.
-- :meth:`.me` will no longer return ``None`` when called in :attr:`.read_only` mode
-  starting in PRAW 8. A :py:class:`DeprecationWarning` will be issued. To switch forward
-  to the PRAW 8 behavior set ``praw8_raise_exception_on_me=True`` in your
+- ``Reddit.user.me`` will no longer return ``None`` when called in :attr:`.read_only`
+  mode starting in PRAW 8. A :py:class:`DeprecationWarning` will be issued. To switch
+  forward to the PRAW 8 behavior set ``praw8_raise_exception_on_me=True`` in your
   :class:`.Reddit` call.
 
 7.1.4 (2021/02/07)